Move more parameter value utils to metabase-lib
