+++ frontend/src/metabase-lib/lib/Question.ts
@@ -31,10 +31,6 @@ import { memoizeClass, sortObject } from "metabase-lib/lib/utils";
 import * as Urls from "metabase/lib/urls";
 import { getCardUiParameters } from "metabase/parameters/utils/cards";
 import {
-  normalizeParameterValue,
-  getParameterValuesBySlug,
-} from "metabase/parameters/utils/parameter-values";
-import {
   DashboardApi,
   CardApi,
   maybeUsePivotEndpoint,
@@ -57,6 +53,10 @@ import { DependentMetadataItem } from "metabase-types/types/Query";
 import { utf8_to_b64url } from "metabase/lib/encoding";
 import { CollectionId } from "metabase-types/api";
 
+import {
+  normalizeParameterValue,
+  getParameterValuesBySlug,
+} from "metabase-lib/lib/parameters/utils/parameter-values";
 import { remapParameterValuesToTemplateTags } from "metabase-lib/lib/parameters/utils/template-tags";
 import { fieldFilterParameterToMBQLFilter } from "metabase-lib/lib/parameters/utils/mbql";
 import { getQuestionVirtualTableId } from "metabase-lib/lib/metadata/utils/saved-questions";
