+++ frontend/src/metabase-lib/lib/queries/utils/card.js
@@ -1,9 +1,8 @@
 import _ from "underscore";
 import { updateIn } from "icepick";
 
-import { normalizeParameterValue } from "metabase/parameters/utils/parameter-values";
-
 import Utils from "metabase/lib/utils";
+import { normalizeParameterValue } from "metabase-lib/lib/parameters/utils/parameter-values";
 import { deriveFieldOperatorFromParameter } from "metabase-lib/lib/parameters/utils/operators";
 import * as Q_DEPRECATED from "metabase-lib/lib/queries/utils"; // legacy
