+++ frontend/src/metabase/dashboard/components/DashCard.jsx
@@ -30,9 +30,9 @@ import { IS_EMBED_PREVIEW } from "metabase/lib/embed";
 
 import { isActionCard } from "metabase/writeback/utils";
 
-import { getParameterValuesBySlug } from "metabase/parameters/utils/parameter-values";
 import Utils from "metabase/lib/utils";
 import { getClickBehaviorDescription } from "metabase/lib/click-behavior";
+import { getParameterValuesBySlug } from "metabase-lib/lib/parameters/utils/parameter-values";
 import DashCardParameterMapper from "./DashCardParameterMapper";
 import { DashCardRoot } from "./DashCard.styled";
