+++ frontend/src/metabase/parameters/utils/parameter-values.js
@@ -1,8 +1,5 @@
 import { getParameterType } from "metabase-lib/lib/parameters/utils/parameter-type";
-import {
-  hasParameterValue,
-  hasDefaultParameterValue,
-} from "metabase-lib/lib/parameters/utils/parameter-values";
+import { hasParameterValue } from "metabase-lib/lib/parameters/utils/parameter-values";
 
 export function getParameterValueFromQueryParams(
   parameter,
@@ -107,54 +104,4 @@ export function getParameterValuesByIdFromQueryParams(
   ]);
 
   return Object.fromEntries(idValuePairs);
-}
-
-function removeNilValuedPairs(pairs) {
-  return pairs.filter(([, value]) => hasParameterValue(value));
-}
-
-function removeUndefaultedNilValuedPairs(pairs) {
-  return pairs.filter(
-    ([parameter, value]) =>
-      hasDefaultParameterValue(parameter) || hasParameterValue(value),
-  );
-}
-
-// when `preserveDefaultedParameters` is true, we don't remove defaulted parameters with nil values
-// so that they can be set in the URL query without a value. Used alongside `getParameterValuesByIdFromQueryParams`
-// with `forcefullyUnsetDefaultedParametersWithEmptyStringValue` set to true.
-export function getParameterValuesBySlug(
-  parameters,
-  parameterValuesById,
-  { preserveDefaultedParameters } = {},
-) {
-  parameters = parameters || [];
-  parameterValuesById = parameterValuesById || {};
-  const parameterValuePairs = parameters.map(parameter => [
-    parameter,
-    hasParameterValue(parameter.value)
-      ? parameter.value
-      : parameterValuesById[parameter.id],
-  ]);
-
-  const transformedPairs = preserveDefaultedParameters
-    ? removeUndefaultedNilValuedPairs(parameterValuePairs)
-    : removeNilValuedPairs(parameterValuePairs);
-
-  const slugValuePairs = transformedPairs.map(([parameter, value]) => [
-    parameter.slug,
-    value,
-  ]);
-
-  return Object.fromEntries(slugValuePairs);
-}
-
-export function normalizeParameterValue(type, value) {
-  const fieldType = getParameterType(type);
-
-  if (["string", "number"].includes(fieldType)) {
-    return value == null ? [] : [].concat(value);
-  } else {
-    return value;
-  }
 }
