+++ frontend/src/metabase/public/containers/PublicQuestion.jsx
@@ -10,10 +10,8 @@ import LoadingAndErrorWrapper from "metabase/components/LoadingAndErrorWrapper";
 import ExplicitSize from "metabase/components/ExplicitSize";
 import title from "metabase/hoc/Title";
 
-import {
-  getParameterValuesBySlug,
-  getParameterValuesByIdFromQueryParams,
-} from "metabase/parameters/utils/parameter-values";
+import { getParameterValuesByIdFromQueryParams } from "metabase/parameters/utils/parameter-values";
+
 import { getCardUiParameters } from "metabase/parameters/utils/cards";
 
 import {
@@ -29,6 +27,7 @@ import { addParamValues, addFields } from "metabase/redux/metadata";
 import { getMetadata } from "metabase/selectors/metadata";
 
 import PublicMode from "metabase/modes/components/modes/PublicMode";
+import { getParameterValuesBySlug } from "metabase-lib/lib/parameters/utils/parameter-values";
 import { getParametersFromCard } from "metabase-lib/lib/parameters/utils/template-tags";
 import { applyParameters } from "metabase-lib/lib/queries/utils/card";
 import EmbedFrame from "../components/EmbedFrame";
